+++ src/components/componentDayTip/index.ts
@@ -1,0 +1,4 @@
+import { register } from '../../shared/functions';
+import ComponentDayTip from './componentDayTip';
+
+register('component-day-tip', ComponentDayTip);